Add more information to the README.md
Add a new section "Pre Requisites" to the README describing the steps
needed to be able to connect to VLC.

Add a new section "Installation" to the README describing how to
install this module.

Add a new section "Usage" to the README showing an example on how to
use this module.

Add a new section "Bugs" to the README describing the current known
bugs.